requirements.txt
# ── Application 1: Traceable Mock Server ──
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0

# ── Application 2: False Positive Minimizer ──
openai>=1.30.0
openai-agents>=0.0.14
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
httpx>=0.27.0
pyyaml>=6.0.0

# ── Evaluation ──
ragas>=0.1.0
datasets>=2.0.0
langchain-openai>=0.1.0

# ── MCP Server ──
mcp>=1.0.0
